Refactor projects table + compare page
